+++ src/rpc/mod.rs
@@ -5,6 +5,7 @@ mod auth_api;
 mod beacon_api;
 mod chain_api;
 mod common_api;
+mod eth_api;
 mod gas_api;
 mod mpool_api;
 mod net_api;
@@ -19,8 +20,19 @@ mod wallet_api;
 use std::sync::Arc;
 
 use crate::rpc_api::{
-    auth_api::*, beacon_api::*, chain_api::*, common_api::*, data_types::RPCState, gas_api::*,
-    mpool_api::*, net_api::*, node_api::NODE_STATUS, state_api::*, sync_api::*, wallet_api::*,
+    auth_api::*,
+    beacon_api::*,
+    chain_api::*,
+    common_api::*,
+    data_types::RPCState,
+    eth_api::{ETH_ACCOUNTS, ETH_BLOCK_NUMBER, ETH_CHAIN_ID},
+    gas_api::*,
+    mpool_api::*,
+    net_api::*,
+    node_api::NODE_STATUS,
+    state_api::*,
+    sync_api::*,
+    wallet_api::*,
 };
 use axum::routing::{get, post};
 use fvm_ipld_blockstore::Blockstore;
@@ -31,8 +43,8 @@ use tracing::info;
 
 use crate::rpc::{
     beacon_api::beacon_get_entry,
-    common_api::{shutdown, start_time, version},
-    rpc_http_handler::rpc_http_handler,
+    common_api::{session, shutdown, start_time, version},
+    rpc_http_handler::{rpc_http_handler, rpc_v0_http_handler},
     rpc_ws_handler::rpc_ws_handler,
     state_api::*,
 };
@@ -120,8 +132,13 @@ where
             .with_method(MINER_GET_BASE_INFO, miner_get_base_info::<DB>)
             .with_method(STATE_MINER_ACTIVE_SECTORS, state_miner_active_sectors::<DB>)
             .with_method(STATE_MINER_FAULTS, state_miner_faults::<DB>)
+            .with_method(STATE_MINER_RECOVERIES, state_miner_recoveries::<DB>)
             .with_method(STATE_MINER_POWER, state_miner_power::<DB>)
             .with_method(STATE_MINER_DEADLINES, state_miner_deadlines::<DB>)
+            .with_method(
+                STATE_MINER_PROVING_DEADLINE,
+                state_miner_proving_deadline::<DB>,
+            )
             .with_method(STATE_GET_RECEIPT, state_get_receipt::<DB>)
             .with_method(STATE_WAIT_MSG, state_wait_msg::<DB>)
             .with_method(STATE_FETCH_ROOT, state_fetch_root::<DB>)
@@ -134,6 +151,7 @@ where
                 state_get_randomness_from_beacon::<DB>,
             )
             .with_method(STATE_READ_STATE, state_read_state::<DB>)
+            .with_method(STATE_CIRCULATING_SUPPLY, state_circulating_supply::<DB>)
             .with_method(STATE_SECTOR_GET_INFO, state_sector_get_info::<DB>)
             .with_method(
                 STATE_VM_CIRCULATING_SUPPLY_INTERNAL,
@@ -146,6 +164,7 @@ where
             .with_method(GAS_ESTIMATE_MESSAGE_GAS, gas_estimate_message_gas::<DB>)
             // Common API
             .with_method(VERSION, move || version(block_delay, forest_version))
+            .with_method(SESSION, session)
             .with_method(SHUTDOWN, move || shutdown(shutdown_send.clone()))
             .with_method(START_TIME, start_time::<DB>)
             // Net API
@@ -156,12 +175,17 @@ where
             .with_method(NET_DISCONNECT, net_api::net_disconnect::<DB>)
             // Node API
             .with_method(NODE_STATUS, node_api::node_status::<DB>)
+            // Eth API
+            .with_method(ETH_ACCOUNTS, eth_api::eth_accounts)
+            .with_method(ETH_BLOCK_NUMBER, eth_api::eth_block_number::<DB>)
+            .with_method(ETH_CHAIN_ID, eth_api::eth_chain_id::<DB>)
             .finish_unwrapped(),
     );
 
     let app = axum::Router::new()
         .route("/rpc/v0", get(rpc_ws_handler))
-        .route("/rpc/v0", post(rpc_http_handler))
+        .route("/rpc/v0", post(rpc_v0_http_handler))
+        .route("/rpc/v1", post(rpc_http_handler))
         .with_state(rpc_server);
 
     info!("Ready for RPC connections");