+++ src/rpc_client/state_ops.rs
@@ -19,7 +19,7 @@ use crate::{
     },
 };
 use cid::Cid;
-use fil_actor_interface::miner::{MinerInfo, MinerPower};
+use fil_actor_interface::miner::{DeadlineInfo, MinerInfo, MinerPower};
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
 use fil_actors_shared::v10::runtime::DomainSeparationTag;
 use libipld_core::ipld::Ipld;
@@ -82,6 +82,10 @@ impl ApiInfo {
         RpcRequest::new(STATE_MINER_FAULTS, (miner, tsk))
     }
 
+    pub fn state_miner_recoveries_req(miner: Address, tsk: TipsetKeys) -> RpcRequest<BitField> {
+        RpcRequest::new(STATE_MINER_RECOVERIES, (miner, tsk))
+    }
+
     pub fn state_miner_power_req(miner: Address, tsk: TipsetKeys) -> RpcRequest<MinerPower> {
         RpcRequest::new(STATE_MINER_POWER, (miner, tsk))
     }
@@ -91,6 +95,13 @@ impl ApiInfo {
         tsk: TipsetKeys,
     ) -> RpcRequest<Vec<ApiDeadline>> {
         RpcRequest::new(STATE_MINER_DEADLINES, (miner, tsk))
+    }
+
+    pub fn state_miner_proving_deadline_req(
+        miner: Address,
+        tsk: TipsetKeys,
+    ) -> RpcRequest<DeadlineInfo> {
+        RpcRequest::new(STATE_MINER_PROVING_DEADLINE, (miner, tsk))
     }
 
     pub fn state_get_randomness_from_tickets_req(